move static kernels to module root
